flask_app/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.3
cycler==0.11.0
Flask==2.0.2
fonttools==4.24.3
idna==3.3
importlib-metadata>=4.0
itsdangerous>=2.0
Jinja2>=3.0
kiwisolver>=1.0
MarkupSafe>=2.0
numpy>=1.00
packaging>=21.0
pandas>=1.0
pyparsing>=3.0
python-dateutil>=2.0
pytz==2022.1
requests>=2.0
six==1.16.0
urllib3==1.26.9
Werkzeug>=2.0
zipp>=3.0.0
uwsgi
datetime
statistics
